+++ src/autosize.js
@@ -3,48 +3,7 @@ const assignedElements = new Map();
 function assign(ta) {
 	if (!ta || !ta.nodeName || ta.nodeName !== 'TEXTAREA' || assignedElements.has(ta)) return;
 
-	let heightOffset = null;
-	let clientWidth = null;
-	let cachedHeight = null;
-
-	function init() {
-		const style = window.getComputedStyle(ta);
-
-		if (style.resize === 'vertical') {
-			ta.style.resize = 'none';
-		} else if (style.resize === 'both') {
-			ta.style.resize = 'horizontal';
-		}
-
-		if (style.boxSizing === 'content-box') {
-			heightOffset = -(parseFloat(style.paddingTop)+parseFloat(style.paddingBottom));
-		} else {
-			heightOffset = parseFloat(style.borderTopWidth)+parseFloat(style.borderBottomWidth);
-		}
-		// Fix when a textarea is not on document body and heightOffset is Not a Number
-		if (isNaN(heightOffset)) {
-			heightOffset = 0;
-		}
-
-		update();
-	}
-
-	function changeOverflow(value) {
-		{
-			// Chrome/Safari-specific fix:
-			// When the textarea y-overflow is hidden, Chrome/Safari do not reflow the text to account for the space
-			// made available by removing the scrollbar. The following forces the necessary text reflow.
-			const width = ta.style.width;
-			ta.style.width = '0px';
-			// Force reflow:
-			/* jshint ignore:start */
-			ta.offsetWidth;
-			/* jshint ignore:end */
-			ta.style.width = width;
-		}
-
-		ta.style.overflowY = value;
-	}
+	let previousHeight = null;
 
 	function cacheScrollTops(el) {
 		const arr = [];
@@ -63,7 +22,11 @@ function assign(ta) {
 		});
 	}
 
-	function resize() {
+	const computed = window.getComputedStyle(ta);
+
+	function update(cachedTextAlign = null) {
+		let initialOverflowY = computed.overflowY;
+
 		if (ta.scrollHeight === 0) {
 			// If the scrollHeight is 0, then the element probably has display:none or is detached from the DOM.
 			return;
@@ -72,67 +35,71 @@ function assign(ta) {
 		// ensure the scrollTop values of parent elements are not modified as a consequence of calculating the textarea height
 		const restoreScrollTops = cacheScrollTops(ta);
 
-		ta.style.height = ''; // this is necessary for getting an accurate scrollHeight on the next line
-		ta.style.height = (ta.scrollHeight+heightOffset)+'px';
+		ta.style.height = ''; // this is necessary for to scrollHeight to accurately reflect situations where the textarea should shrink
 
-		// used to check if an update is actually necessary on window.resize
-		clientWidth = ta.clientWidth;
+		// disallow vertical resizing
+		if (computed.resize === 'vertical') {
+			ta.style.resize = 'none';
+		} else if (computed.resize === 'both') {
+			ta.style.resize = 'horizontal';
+		}
+
+		let newHeight;
+
+		if (computed.boxSizing === 'content-box') {
+			newHeight = ta.scrollHeight - (parseFloat(computed.paddingTop)+parseFloat(computed.paddingBottom));
+		} else {
+			newHeight = ta.scrollHeight + parseFloat(computed.borderTopWidth)+parseFloat(computed.borderBottomWidth);
+		}
+
+		if (computed.maxHeight !== 'none' && newHeight > parseFloat(computed.maxHeight)) {
+			if (computed.overflowY === 'hidden') {
+				ta.style.overflow = 'scroll';
+			}
+			newHeight = parseFloat(computed.maxHeight);
+		} else if (computed.overflowY !== 'hidden') {
+			ta.style.overflow = 'hidden';
+		}
+
+		ta.style.height = newHeight+'px';
+
+		if (cachedTextAlign) {
+			ta.style.textAlign = cachedTextAlign;
+		}
 
 		restoreScrollTops();
-	}
 
-	function update() {
-		resize();
-
-		const styleHeight = parseFloat(ta.style.height);
-		const computed = window.getComputedStyle(ta);
-
-		var actualHeight = parseFloat(computed.height);
-
-		// The actual height not matching the style height (set via the resize method) indicates that
-		// the max-height has been exceeded, in which case the overflow should be allowed.
-		if (actualHeight < styleHeight) {
-			if (computed.overflowY === 'hidden') {
-				changeOverflow('scroll');
-				resize();
-				actualHeight = parseFloat(window.getComputedStyle(ta).height);
-			}
-		} else {
-			// Normally keep overflow set to hidden, to avoid flash of scrollbar as the textarea expands.
-			if (computed.overflowY !== 'hidden') {
-				changeOverflow('hidden');
-				resize();
-				actualHeight = parseFloat(window.getComputedStyle(ta).height);
-			}
-		}
-
-		if (cachedHeight !== actualHeight) {
-			cachedHeight = actualHeight;
+		if (previousHeight !== newHeight) {
 			ta.dispatchEvent(new Event('autosize:resized', {bubbles: true}));
+			previousHeight = newHeight;
+		}
+
+		if (initialOverflowY !== computed.overflow && !cachedTextAlign) {
+			const textAlign = computed.textAlign;
+
+			if (computed.overflow === 'hidden') {
+				// Webkit fails to reflow text after overflow is hidden,
+				// even if hiding overflow would allow text to fit more compactly.
+				// The following is intended to force the necessary text reflow.
+				ta.style.textAlign = textAlign === 'start' ? 'end' : 'start';
+			}
+
+			update(textAlign);
 		}
 	}
-
-	const pageResize = () => {
-		if (ta.clientWidth !== clientWidth) {
-			update();
-		}
-	};
 
 	const destroy = (style => {
-		window.removeEventListener('resize', pageResize, false);
+		window.removeEventListener('resize', update, false);
 		ta.removeEventListener('input', update, false);
 		ta.removeEventListener('keyup', update, false);
 		ta.removeEventListener('autosize:destroy', destroy, false);
 		ta.removeEventListener('autosize:update', update, false);
-
-		Object.keys(style).forEach(key => {
-			ta.style[key] = style[key];
-		});
-
+		Object.keys(style).forEach(key => ta.style[key] = style[key]);
 		assignedElements.delete(ta);
 	}).bind(ta, {
 		height: ta.style.height,
 		resize: ta.style.resize,
+		textAlign: ta.style.textAlign,
 		overflowY: ta.style.overflowY,
 		overflowX: ta.style.overflowX,
 		wordWrap: ta.style.wordWrap,
@@ -140,7 +107,7 @@ function assign(ta) {
 
 	ta.addEventListener('autosize:destroy', destroy, false);
 
-	window.addEventListener('resize', pageResize, false);
+	window.addEventListener('resize', update, false);
 	ta.addEventListener('input', update, false);
 	ta.addEventListener('autosize:update', update, false);
 	ta.style.overflowX = 'hidden';
@@ -151,7 +118,7 @@ function assign(ta) {
 		update,
 	});
 
-	init();
+	update();
 }
 
 function destroy(ta) {